Check users permission in dm thread before sending message
